reorganized file structure, components for icons, begin css modules
