+++ src/components/MobileSocialLinks/index.tsx
@@ -1,97 +1,20 @@
 import React from 'react';
+import classes from './style.module.css';
+import GithubIcon from '@site/src/components/icons/GithubIcon';
+import DiscordIcon from '@site/src/components/icons/DiscordIcon';
+import TwitterIcon from '@site/src/components/icons/TwitterIcon';
 
 export default function MobileSocialLinks() {
   return (
-    <div style={{ display: 'flex', marginLeft: '15px' }}>
-      <a style={{ margin: '5px' }} href="https://github.com/Kava-Labs">
-        <svg
-          width="20"
-          height="19"
-          viewBox="0 0 20 19"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <g clip-path="url(#clip0_7202_77407)">
-            <path
-              fill-rule="evenodd"
-              clip-rule="evenodd"
-              d="M9.85915 0C4.40723 0 0 4.35434 0 9.74084C0 14.0468 2.82389 17.6915 6.74142 18.9817C7.23112 19.0785 7.41069 18.7721 7.41069 18.514C7.41069 18.2883 7.39435 17.5141 7.39435 16.7078C4.65208 17.2884 4.08077 15.5466 4.08077 15.5466C3.64005 14.4177 2.98712 14.1274 2.98712 14.1274C2.08935 13.5307 3.05242 13.5307 3.05242 13.5307C4.04812 13.5952 4.57046 14.5306 4.57046 14.5306C5.45192 16.0143 6.872 15.595 7.44331 15.337C7.52492 14.708 7.78612 14.2726 8.06362 14.0307C5.87631 13.8049 3.57475 12.9663 3.57475 9.22477C3.57475 8.16035 3.9665 7.2895 4.58677 6.61215C4.48885 6.37024 4.14608 5.37035 4.68473 4.0318C4.68473 4.0318 5.51719 3.77376 7.39435 5.03169C8.17785 4.82205 9.02666 4.70915 9.85915 4.70915C10.6916 4.70915 11.5404 4.82205 12.3239 5.03169C14.2011 3.77376 15.0335 4.0318 15.0335 4.0318C15.5722 5.37035 15.2294 6.37024 15.1315 6.61215C15.7681 7.2895 16.1435 8.16035 16.1435 9.22477C16.1435 12.9663 13.842 13.7888 11.6383 14.0307C11.9975 14.3371 12.3076 14.9177 12.3076 15.8369C12.3076 17.1432 12.2913 18.1915 12.2913 18.514C12.2913 18.7721 12.4708 19.0785 12.9605 18.9817C16.8781 17.6915 19.702 14.0468 19.702 9.74084C19.7183 4.35434 15.2947 0 9.85915 0Z"
-              fill="#1E1E1E"
-            />
-            <path
-              d="M3.73766 13.9822C3.72134 14.0306 3.63972 14.0468 3.57443 14.0145C3.50914 13.9822 3.46017 13.9177 3.49281 13.8693C3.50914 13.821 3.59075 13.8049 3.65604 13.8371C3.72134 13.8693 3.75398 13.9339 3.73766 13.9822Z"
-              fill="#1E1E1E"
-            />
-            <path
-              d="M4.13006 14.4181C4.0811 14.4664 3.98313 14.4342 3.93417 14.3697C3.86886 14.3052 3.85256 14.2084 3.90152 14.16C3.95048 14.1116 4.0321 14.1439 4.0974 14.2084C4.16271 14.289 4.17902 14.3858 4.13006 14.4181Z"
-              fill="#1E1E1E"
-            />
-            <path
-              d="M4.52286 14.9819C4.45759 15.0304 4.35963 14.9819 4.31067 14.9013C4.2454 14.8207 4.2454 14.7078 4.31067 14.6756C4.37598 14.6272 4.4739 14.6756 4.52286 14.7562C4.58817 14.8368 4.58817 14.9336 4.52286 14.9819Z"
-              fill="#1E1E1E"
-            />
-            <path
-              d="M5.0597 15.5308C5.01074 15.5953 4.89647 15.5791 4.79855 15.4985C4.71693 15.4179 4.68428 15.305 4.74959 15.2566C4.79855 15.1921 4.91282 15.2082 5.01074 15.2889C5.09236 15.3534 5.1087 15.4663 5.0597 15.5308Z"
-              fill="#1E1E1E"
-            />
-            <path
-              d="M5.79532 15.8368C5.77901 15.9174 5.66474 15.9497 5.55047 15.9174C5.43624 15.8851 5.37094 15.7884 5.38724 15.7239C5.40359 15.6432 5.51786 15.611 5.63209 15.6432C5.74636 15.6755 5.81167 15.7561 5.79532 15.8368Z"
-              fill="#1E1E1E"
-            />
-            <path
-              d="M6.59239 15.901C6.59239 15.9816 6.49443 16.0461 6.3802 16.0461C6.26593 16.0461 6.16797 15.9816 6.16797 15.901C6.16797 15.8203 6.26593 15.7559 6.3802 15.7559C6.49443 15.7559 6.59239 15.8203 6.59239 15.901Z"
-              fill="#1E1E1E"
-            />
-            <path
-              d="M7.3457 15.7722C7.36201 15.8529 7.28039 15.9335 7.16616 15.9496C7.05189 15.9658 6.95393 15.9174 6.93762 15.8367C6.92132 15.7561 7.00293 15.6754 7.11716 15.6593C7.23143 15.6432 7.32939 15.6916 7.3457 15.7722Z"
-              fill="#1E1E1E"
-            />
-          </g>
-          <defs>
-            <clipPath id="clip0_7202_77407">
-              <rect width="20" height="19" fill="white" />
-            </clipPath>
-          </defs>
-        </svg>
+    <div className={classes.mobileSocialLinksWrapper}>
+      <a className={classes.linkWrapper} href="https://github.com/Kava-Labs">
+        <GithubIcon />
       </a>
-      <a style={{ margin: '5px' }} href="https://discord.gg/JJYnuCx">
-        <svg
-          width="17"
-          height="20"
-          viewBox="0 0 17 20"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <g clip-path="url(#clip0_7202_77404)">
-            <path
-              d="M6.74097 8.39076C6.5237 8.37167 6.3059 8.42056 6.11609 8.53105C5.92628 8.64153 5.7733 8.80844 5.6772 9.00993C5.5811 9.21142 5.54636 9.43808 5.57752 9.66024C5.60867 9.8824 5.70428 10.0897 5.85182 10.255C5.99936 10.4203 6.19195 10.5358 6.40437 10.5865C6.61679 10.6372 6.83913 10.6207 7.04228 10.5391C7.24542 10.4576 7.41991 10.3147 7.54287 10.1294C7.66583 9.94401 7.73154 9.72473 7.7314 9.50027C7.74299 9.21794 7.64515 8.94244 7.45941 8.73436C7.27367 8.52629 7.01524 8.40269 6.74097 8.39076ZM10.287 8.39076C10.0697 8.37167 9.85191 8.42056 9.6621 8.53105C9.47229 8.64153 9.31931 8.80844 9.22321 9.00993C9.12711 9.21142 9.09236 9.43808 9.12352 9.66024C9.15468 9.8824 9.25028 10.0897 9.39782 10.255C9.54536 10.4203 9.73796 10.5358 9.95037 10.5865C10.1628 10.6372 10.3851 10.6207 10.5883 10.5391C10.7914 10.4576 10.9659 10.3147 11.0889 10.1294C11.2118 9.94401 11.2775 9.72473 11.2774 9.50027C11.289 9.21794 11.1912 8.94244 11.0054 8.73436C10.8197 8.52629 10.5612 8.40269 10.287 8.39076Z"
-              fill="#1E1E1E"
-            />
-            <path
-              d="M15.0089 0H1.98923C1.46016 0.00202782 0.953512 0.220122 0.580539 0.606387C0.207567 0.992653 -0.00123032 1.51551 5.4544e-06 2.06012V15.5801C-0.00122947 16.1251 0.207821 16.6482 0.581196 17.0345C0.95457 17.4209 1.4617 17.6387 1.99109 17.6402H13.0076L12.4924 15.7907L13.7358 16.9807L14.9113 18.1007L17 20.001V2.06012C17.0012 1.51518 16.7922 0.992034 16.4188 0.605714C16.0454 0.219394 15.5383 0.00152142 15.0089 0ZM11.2592 13.0595C11.2592 13.0595 10.9096 12.6297 10.6185 12.2497C11.3226 12.0788 11.9455 11.6573 12.3771 11.0597C12.0279 11.299 11.6535 11.4968 11.2611 11.6494C10.8091 11.848 10.3363 11.9922 9.8522 12.0793C9.02037 12.237 8.16727 12.2338 7.3366 12.0697C6.84872 11.9715 6.37058 11.8277 5.90816 11.6399C5.66434 11.5435 5.42726 11.4299 5.19858 11.3C5.16975 11.2799 5.13999 11.2703 5.11116 11.2502C5.0965 11.2429 5.08326 11.2328 5.0721 11.2206C4.89727 11.1248 4.80055 11.0502 4.80055 11.0502C5.21623 11.6356 5.81755 12.0531 6.50055 12.2305C6.20947 12.6134 5.84957 13.0605 5.84957 13.0605C5.27187 13.0771 4.69907 12.9471 4.1817 12.682C3.66433 12.4169 3.2183 12.0248 2.88294 11.5403C2.91326 9.51334 3.39024 7.51991 4.2779 5.71032C5.05984 5.0768 6.01358 4.70885 7.00739 4.65728L7.10039 4.77695C6.16934 5.01771 5.30105 5.46608 4.5569 6.09037C4.5569 6.09037 4.77079 5.97071 5.12976 5.80318C5.82799 5.47511 6.57214 5.26253 7.33474 5.17327C7.38925 5.16163 7.44464 5.1549 7.50028 5.15317C8.15124 5.06583 8.81001 5.05908 9.46253 5.13306C10.4885 5.2529 11.4817 5.5781 12.3864 6.09037C11.6768 5.49555 10.8527 5.06256 9.96844 4.82003L10.1042 4.66016C11.098 4.71172 12.0518 5.07967 12.8337 5.71319C13.7214 7.52278 14.1983 9.51621 14.2287 11.5432C13.8909 12.0255 13.4439 12.4157 12.9264 12.6799C12.409 12.9442 11.8367 13.0745 11.2592 13.0595Z"
-              fill="#1E1E1E"
-            />
-          </g>
-          <defs>
-            <clipPath id="clip0_7202_77404">
-              <rect width="17" height="20" fill="white" />
-            </clipPath>
-          </defs>
-        </svg>
+      <a className={classes.linkWrapper} href="https://discord.gg/JJYnuCx">
+        <DiscordIcon />
       </a>
-      <a style={{ margin: '5px' }} href="https://twitter.com/kava_labs">
-        <svg
-          width="20"
-          height="16"
-          viewBox="0 0 20 16"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M19.9378 1.48429L16.9293 15.0357C16.7206 15.997 16.1125 16.228 15.2621 15.7809L10.6772 12.5469L8.45972 14.5886C8.35267 14.7247 8.21444 14.8358 8.05555 14.9133C7.89667 14.9907 7.72132 15.0326 7.54289 15.0357C6.94261 15.0357 7.04265 14.8211 6.83943 14.2809L5.25588 9.36283L0.715468 8.00888C-0.267805 7.73317 -0.276403 7.08712 0.931976 6.61544L18.6551 0.102028C19.4641 -0.239999 20.2387 0.292788 19.9307 1.47684L19.9378 1.48429Z"
-            fill="#1E1E1E"
-          />
-        </svg>
+      <a className={classes.linkWrapper} href="https://twitter.com/kava_labs">
+        <TwitterIcon />
       </a>
     </div>
   );
