+++ src/components/MobileSocialLinks/style.module.css
@@ -1,0 +1,8 @@
+.linkWrapper {
+  margin: 5px;
+}
+
+.mobileSocialLinksWrapper {
+  display: flex;
+  marginleft: 15px !important;
+}
