+++ src/components/icons/DiscordIcon/index.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+
+export default function DiscordIcon() {
+  return (
+    <svg
+      width="17"
+      height="20"
+      viewBox="0 0 17 20"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g clip-path="url(#clip0_7202_77404)">
+        <path
+          d="M6.74097 8.39076C6.5237 8.37167 6.3059 8.42056 6.11609 8.53105C5.92628 8.64153 5.7733 8.80844 5.6772 9.00993C5.5811 9.21142 5.54636 9.43808 5.57752 9.66024C5.60867 9.8824 5.70428 10.0897 5.85182 10.255C5.99936 10.4203 6.19195 10.5358 6.40437 10.5865C6.61679 10.6372 6.83913 10.6207 7.04228 10.5391C7.24542 10.4576 7.41991 10.3147 7.54287 10.1294C7.66583 9.94401 7.73154 9.72473 7.7314 9.50027C7.74299 9.21794 7.64515 8.94244 7.45941 8.73436C7.27367 8.52629 7.01524 8.40269 6.74097 8.39076ZM10.287 8.39076C10.0697 8.37167 9.85191 8.42056 9.6621 8.53105C9.47229 8.64153 9.31931 8.80844 9.22321 9.00993C9.12711 9.21142 9.09236 9.43808 9.12352 9.66024C9.15468 9.8824 9.25028 10.0897 9.39782 10.255C9.54536 10.4203 9.73796 10.5358 9.95037 10.5865C10.1628 10.6372 10.3851 10.6207 10.5883 10.5391C10.7914 10.4576 10.9659 10.3147 11.0889 10.1294C11.2118 9.94401 11.2775 9.72473 11.2774 9.50027C11.289 9.21794 11.1912 8.94244 11.0054 8.73436C10.8197 8.52629 10.5612 8.40269 10.287 8.39076Z"
+          fill="#1E1E1E"
+        />
+        <path
+          d="M15.0089 0H1.98923C1.46016 0.00202782 0.953512 0.220122 0.580539 0.606387C0.207567 0.992653 -0.00123032 1.51551 5.4544e-06 2.06012V15.5801C-0.00122947 16.1251 0.207821 16.6482 0.581196 17.0345C0.95457 17.4209 1.4617 17.6387 1.99109 17.6402H13.0076L12.4924 15.7907L13.7358 16.9807L14.9113 18.1007L17 20.001V2.06012C17.0012 1.51518 16.7922 0.992034 16.4188 0.605714C16.0454 0.219394 15.5383 0.00152142 15.0089 0ZM11.2592 13.0595C11.2592 13.0595 10.9096 12.6297 10.6185 12.2497C11.3226 12.0788 11.9455 11.6573 12.3771 11.0597C12.0279 11.299 11.6535 11.4968 11.2611 11.6494C10.8091 11.848 10.3363 11.9922 9.8522 12.0793C9.02037 12.237 8.16727 12.2338 7.3366 12.0697C6.84872 11.9715 6.37058 11.8277 5.90816 11.6399C5.66434 11.5435 5.42726 11.4299 5.19858 11.3C5.16975 11.2799 5.13999 11.2703 5.11116 11.2502C5.0965 11.2429 5.08326 11.2328 5.0721 11.2206C4.89727 11.1248 4.80055 11.0502 4.80055 11.0502C5.21623 11.6356 5.81755 12.0531 6.50055 12.2305C6.20947 12.6134 5.84957 13.0605 5.84957 13.0605C5.27187 13.0771 4.69907 12.9471 4.1817 12.682C3.66433 12.4169 3.2183 12.0248 2.88294 11.5403C2.91326 9.51334 3.39024 7.51991 4.2779 5.71032C5.05984 5.0768 6.01358 4.70885 7.00739 4.65728L7.10039 4.77695C6.16934 5.01771 5.30105 5.46608 4.5569 6.09037C4.5569 6.09037 4.77079 5.97071 5.12976 5.80318C5.82799 5.47511 6.57214 5.26253 7.33474 5.17327C7.38925 5.16163 7.44464 5.1549 7.50028 5.15317C8.15124 5.06583 8.81001 5.05908 9.46253 5.13306C10.4885 5.2529 11.4817 5.5781 12.3864 6.09037C11.6768 5.49555 10.8527 5.06256 9.96844 4.82003L10.1042 4.66016C11.098 4.71172 12.0518 5.07967 12.8337 5.71319C13.7214 7.52278 14.1983 9.51621 14.2287 11.5432C13.8909 12.0255 13.4439 12.4157 12.9264 12.6799C12.409 12.9442 11.8367 13.0745 11.2592 13.0595Z"
+          fill="#1E1E1E"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_7202_77404">
+          <rect width="17" height="20" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
