+++ src/components/icons/GithubIcon/index.tsx
@@ -1,0 +1,55 @@
+import React from 'react';
+
+export default function GithubIcon() {
+  return (
+    <svg
+      width="20"
+      height="19"
+      viewBox="0 0 20 19"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g clip-path="url(#clip0_7202_77407)">
+        <path
+          fill-rule="evenodd"
+          clip-rule="evenodd"
+          d="M9.85915 0C4.40723 0 0 4.35434 0 9.74084C0 14.0468 2.82389 17.6915 6.74142 18.9817C7.23112 19.0785 7.41069 18.7721 7.41069 18.514C7.41069 18.2883 7.39435 17.5141 7.39435 16.7078C4.65208 17.2884 4.08077 15.5466 4.08077 15.5466C3.64005 14.4177 2.98712 14.1274 2.98712 14.1274C2.08935 13.5307 3.05242 13.5307 3.05242 13.5307C4.04812 13.5952 4.57046 14.5306 4.57046 14.5306C5.45192 16.0143 6.872 15.595 7.44331 15.337C7.52492 14.708 7.78612 14.2726 8.06362 14.0307C5.87631 13.8049 3.57475 12.9663 3.57475 9.22477C3.57475 8.16035 3.9665 7.2895 4.58677 6.61215C4.48885 6.37024 4.14608 5.37035 4.68473 4.0318C4.68473 4.0318 5.51719 3.77376 7.39435 5.03169C8.17785 4.82205 9.02666 4.70915 9.85915 4.70915C10.6916 4.70915 11.5404 4.82205 12.3239 5.03169C14.2011 3.77376 15.0335 4.0318 15.0335 4.0318C15.5722 5.37035 15.2294 6.37024 15.1315 6.61215C15.7681 7.2895 16.1435 8.16035 16.1435 9.22477C16.1435 12.9663 13.842 13.7888 11.6383 14.0307C11.9975 14.3371 12.3076 14.9177 12.3076 15.8369C12.3076 17.1432 12.2913 18.1915 12.2913 18.514C12.2913 18.7721 12.4708 19.0785 12.9605 18.9817C16.8781 17.6915 19.702 14.0468 19.702 9.74084C19.7183 4.35434 15.2947 0 9.85915 0Z"
+          fill="#1E1E1E"
+        />
+        <path
+          d="M3.73766 13.9822C3.72134 14.0306 3.63972 14.0468 3.57443 14.0145C3.50914 13.9822 3.46017 13.9177 3.49281 13.8693C3.50914 13.821 3.59075 13.8049 3.65604 13.8371C3.72134 13.8693 3.75398 13.9339 3.73766 13.9822Z"
+          fill="#1E1E1E"
+        />
+        <path
+          d="M4.13006 14.4181C4.0811 14.4664 3.98313 14.4342 3.93417 14.3697C3.86886 14.3052 3.85256 14.2084 3.90152 14.16C3.95048 14.1116 4.0321 14.1439 4.0974 14.2084C4.16271 14.289 4.17902 14.3858 4.13006 14.4181Z"
+          fill="#1E1E1E"
+        />
+        <path
+          d="M4.52286 14.9819C4.45759 15.0304 4.35963 14.9819 4.31067 14.9013C4.2454 14.8207 4.2454 14.7078 4.31067 14.6756C4.37598 14.6272 4.4739 14.6756 4.52286 14.7562C4.58817 14.8368 4.58817 14.9336 4.52286 14.9819Z"
+          fill="#1E1E1E"
+        />
+        <path
+          d="M5.0597 15.5308C5.01074 15.5953 4.89647 15.5791 4.79855 15.4985C4.71693 15.4179 4.68428 15.305 4.74959 15.2566C4.79855 15.1921 4.91282 15.2082 5.01074 15.2889C5.09236 15.3534 5.1087 15.4663 5.0597 15.5308Z"
+          fill="#1E1E1E"
+        />
+        <path
+          d="M5.79532 15.8368C5.77901 15.9174 5.66474 15.9497 5.55047 15.9174C5.43624 15.8851 5.37094 15.7884 5.38724 15.7239C5.40359 15.6432 5.51786 15.611 5.63209 15.6432C5.74636 15.6755 5.81167 15.7561 5.79532 15.8368Z"
+          fill="#1E1E1E"
+        />
+        <path
+          d="M6.59239 15.901C6.59239 15.9816 6.49443 16.0461 6.3802 16.0461C6.26593 16.0461 6.16797 15.9816 6.16797 15.901C6.16797 15.8203 6.26593 15.7559 6.3802 15.7559C6.49443 15.7559 6.59239 15.8203 6.59239 15.901Z"
+          fill="#1E1E1E"
+        />
+        <path
+          d="M7.3457 15.7722C7.36201 15.8529 7.28039 15.9335 7.16616 15.9496C7.05189 15.9658 6.95393 15.9174 6.93762 15.8367C6.92132 15.7561 7.00293 15.6754 7.11716 15.6593C7.23143 15.6432 7.32939 15.6916 7.3457 15.7722Z"
+          fill="#1E1E1E"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_7202_77407">
+          <rect width="20" height="19" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
