+++ src/components/icons/LightModeIcon/index.tsx
