+++ src/components/icons/TwitterIcon/index.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+export default function TwitterIcon() {
+  return (
+    <svg
+      width="20"
+      height="16"
+      viewBox="0 0 20 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M19.9378 1.48429L16.9293 15.0357C16.7206 15.997 16.1125 16.228 15.2621 15.7809L10.6772 12.5469L8.45972 14.5886C8.35267 14.7247 8.21444 14.8358 8.05555 14.9133C7.89667 14.9907 7.72132 15.0326 7.54289 15.0357C6.94261 15.0357 7.04265 14.8211 6.83943 14.2809L5.25588 9.36283L0.715468 8.00888C-0.267805 7.73317 -0.276403 7.08712 0.931976 6.61544L18.6551 0.102028C19.4641 -0.239999 20.2387 0.292788 19.9307 1.47684L19.9378 1.48429Z"
+        fill="#1E1E1E"
+      />
+    </svg>
+  );
+}
